Terminals: Bug 435150 - Support CDT 8.3 minimal dependency via p2.inf
This change makes sure that the Terminal can depend on a minimal
subset of CDT functionality, imported by import-package in MANIFEST.MF
plus fragment dependencies in p2.inf.

The org.eclipse.cdt.core.native feature is pulled in if it's available,
but we can also work on top of CDT 8.3 which didn't have the "native"
feature yet.

Cherry-pick from master, resolved conflicts:
	target_explorer/features/org.eclipse.tcf.te.terminals.feature/feature.xml

Change-Id: Ibd2dbe1b971f52195aa28fdcdb7e1048ad1ed3cf
Signed-off-by: Martin Oberhuber <martin.oberhuber@windriver.com>